GRPC_emulated/GRPC-client/requirements.txt

grpcio
grpcio-tools
motor
python-dotenv
uvloop
numpy